backend/requirements.txt
fastapi
uvicorn
face_recognition
numpy
Pillow
python-multipart
opencv-python
